+++ types/cryptoUtils.d.ts
@@ -1,3 +1,4 @@
+import { mnemonicToSeed } from 'bip39';
 interface Keys {
     pk: string;
     pkh: string;
@@ -31,9 +32,10 @@ interface Signed {
 export declare const extractKeys: (sk: string, passphrase?: string) => Promise<Keys>;
 /**
  * @description Generate a mnemonic
- * @returns {string} The 15 word generated mnemonic
+ * @param {number} [numberOfWords=15] The number of words to include in the mnemonic
+ * @returns {string} The generated mnemonic
  */
-export declare const generateMnemonic: () => string;
+export declare const generateMnemonic: (numberOfWords?: number) => string;
 /**
  * @description Check the validity of a tezos implicit address (tz1...)
  * @param {string} address The address to check
@@ -43,13 +45,18 @@ export declare const checkAddress: (address: string) => boolean;
 /**
  * @description Generate a new key pair given a mnemonic and passphrase
  * @param {string} mnemonic The mnemonic seed
- * @param {string} passphrase The passphrase used to encrypt the seed
+ * @param {string} [passphrase] The passphrase used to salt the seed
+ * @param {string} [derivationPath='m/44h/1729h/0h/0h'] Derivation path if generating keys for an HD account
  * @returns {Promise} The generated key pair
  * @example
  * cryptoUtils.generateKeys('raw peace visual boil prefer rebel anchor right elegant side gossip enroll force salmon between', 'my_password_123')
  *   .then(({ mnemonic, passphrase, sk, pk, pkh }) => console.log(mnemonic, passphrase, sk, pk, pkh));
+ *
+ * // Or generate keys given an HD derivartion path
+ * cryptoUtils.generateKeys('gym exact clown can answer hope sample mirror knife twenty powder super imitate lion churn almost shed chalk dust civil gadget pyramid helmet trade', undefined, 'm/44h/1729h/0h/0h')
+ *   .then(({ mnemonic, passphrase, sk, pk, pkh }) => console.log(mnemonic, passphrase, sk, pk, pkh));
  */
-export declare const generateKeys: (mnemonic: string, passphrase?: string) => Promise<KeysMnemonicPassphrase>;
+export declare const generateKeys: (mnemonic: string, passphrase?: string, derivationPath?: string) => Promise<KeysMnemonicPassphrase>;
 /**
  * @description Encrypts a secret key with a passphrase
  * @param {string} key The secret key
@@ -86,11 +93,12 @@ export declare const sign: (bytes: string, sk: string, magicBytes?: Uint8Array, 
  */
 export declare const verify: (bytes: string, sig: string, pk: string) => Promise<boolean>;
 declare const _default: {
-    extractKeys: (sk: string, passphrase?: string) => Promise<Keys>;
+    mnemonicToSeed: typeof mnemonicToSeed;
+    extractKeys: (sk: string, passphrase?: string | undefined) => Promise<Keys>;
     encryptSecretKey: (key: string, passphrase: string, salt?: Uint8Array) => string;
-    generateKeys: (mnemonic: string, passphrase?: string | undefined) => Promise<KeysMnemonicPassphrase>;
+    generateKeys: (mnemonic: string, passphrase?: string | undefined, derivationPath?: string | undefined) => Promise<KeysMnemonicPassphrase>;
     checkAddress: (address: string) => boolean;
-    generateMnemonic: () => string;
+    generateMnemonic: (numberOfWords?: number) => string;
     sign: (bytes: string, sk: string, magicBytes?: Uint8Array | undefined, password?: string) => Promise<Signed>;
     verify: (bytes: string, sig: string, pk: string) => Promise<boolean>;
 };